Updated the GUI with an early version of image display code after FelixSim runs. This uses a Bin2Tiff program that converts the 32bit binaries to 8 bit .tif files
